SY-4740: Fold the keep-alive specs into streamer.spec.ts
